refactor: TodoItem dom객체 표현 부분 todoTemplate로 분리
